aggiunta classi di test.

+++ src/main/java/it/plansoft/chinookjpa/controller/BaseCrudController.java
@@ -1,6 +1,7 @@
 package it.plansoft.chinookjpa.controller;/* ggrosso created on 05/03/2021 inside the package - it.plansoft.chinook.controller */
 
 import it.plansoft.chinookjpa.controller.interfaces.ICrudController;
+import it.plansoft.chinookjpa.model.BaseId;
 import it.plansoft.chinookjpa.service.BaseCrudService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -14,10 +15,9 @@ import java.util.Optional;
 public class BaseCrudController<
         SERVICE extends BaseCrudService<REPOSITORY, MODEL, ID>,
         REPOSITORY extends JpaRepository<MODEL, ID>,
-        MODEL,
+        MODEL extends BaseId<ID>,
         ID>
-        implements ICrudController<MODEL, ID>
-{
+        implements ICrudController<MODEL, ID> {
 
     protected SERVICE service;
 
@@ -58,14 +58,24 @@ public class BaseCrudController<
 
     @Override
     @DeleteMapping("/")
-    public void delete(MODEL model) {
-        service.delete(model);
+    public ResponseEntity<Optional<MODEL>> delete(MODEL model) {
+        Optional<MODEL> m = service.findById(model.getId());
+
+        if (m != null && m.get() != null)
+            service.delete(model);
+
+        return ResponseEntity.ok(m);
     }
 
     @Override
     @PostMapping("/{id}")
-    public void deleteById(@PathVariable ID id) {
-        service.deleteById(id);
+    public ResponseEntity<Optional<MODEL>> deleteById(@PathVariable ID id) {
+        Optional<MODEL> m = service.findById(id);
+
+        if (m != null && m.get() != null)
+            service.deleteById(id);
+
+        return ResponseEntity.ok(m);
     }
 
     @Override
